Update login and register page
