+++ src/routes.tsx
@@ -8,10 +8,6 @@ import MinimalLayout from 'layout/MinimalLayout';
 // landing routing
 const LandingPage = Loadable(lazy(() => import('views/landing')));
 
-const LoginPage = Loadable(lazy(() => import('views/login')));
-
-const SignupPage = Loadable(lazy(() => import('views/signup')));
-
 // dashboard routing
 const Dashboard = Loadable(lazy(() => import('views/dashboard')));
 
@@ -68,11 +64,7 @@ const Routes = () => {
         },
         {
           path: '',
-          element: <LoginPage />
-        },
-        {
-          path: 'signup',
-          element: <SignupPage />
+          element: <AuthLogin />
         },
         {
           path: '*',
